refactor: move specific helper inside test
