--- src/testHelper.ts
@@ -1,13 +1,0 @@
-export const rootStyle = {
-  display: 'flex',
-  justifyContent: 'center',
-};
-export const rowStyle = {
-  margin: '200px 0',
-  display: 'flex',
-  justifyContent: 'space-between',
-};
-export const boxStyle = {
-  padding: '10px',
-  border: '1px solid black',
-};
